refactor keeper registry settings
