+++ contracts/ethereum/generated.go
@@ -8,3 +8,11 @@ import (
 type AbigenLog interface {
 	Topic() common.Hash
 }
+
+type KeeperRegistryVersion int32
+
+const (
+	RegistryVersion_1_0 KeeperRegistryVersion = iota
+	RegistryVersion_1_1
+	RegistryVersion_1_2
+)
