+++ testsetups/keeper.go
@@ -12,6 +12,7 @@ import (
 	"github.com/smartcontractkit/chainlink-testing-framework/blockchain"
 	"github.com/smartcontractkit/chainlink-testing-framework/client"
 	"github.com/smartcontractkit/chainlink-testing-framework/contracts"
+	"github.com/smartcontractkit/chainlink-testing-framework/contracts/ethereum"
 	"github.com/smartcontractkit/chainlink-testing-framework/testreporters"
 	"github.com/smartcontractkit/helmenv/environment"
 )
@@ -79,6 +80,7 @@ func (k *KeeperBlockTimeTest) Setup(env *environment.Environment) {
 	Expect(err).ShouldNot(HaveOccurred(), "Failed waiting for LINK Contract deployment")
 
 	k.keeperRegistry, k.keeperConsumerContracts = actions.DeployPerformanceKeeperContracts(
+		ethereum.RegistryVersion_1_1,
 		inputs.NumberOfContracts,
 		linkToken,
 		contractDeployer,
